requirements_word_converter.txt
python-docx>=0.8.11
requests>=2.31.0
cairosvg>=2.7.0
